add termplot with fixed config.yaml (#129)
This adds the termplot plugin to the list again because the previous PR
was reverted due to a CI error.

This time I have added "override" config like some other non-Rust
plugins.

The previous PR was here: https://github.com/nushell/awesome-nu/pull/127

I believe the issue was that the CI script assumes the presence of a
Cargo.toml file. For non-Rust plugins like this one, there is no such
Cargo.toml file. Instead, if the other plugins are a guide, one must use
the "override" settings to put the values that would otherwise be taken
from the Cargo.toml file.