remove packer.nu (#139)
packer.nu is still maintained, but i (as the maintainer) don't want to
grow its adoption.

the package-design is flawed, the package-count is tiny, and noone is
expanding it.
new people still find it (presumable through this), but would probably
have a way better time with `nupm` or creating something new.